Thunderdome v2 ww02 release (weekly-09 build)
It includes the prelimenary support of BLE central
It also includes several fixes to improve the stability

+++ v2.0.0/device_software/intel_iq_sdk/wearable_device_sw/bsp/src/os/zephyr/epoch_time.c
@@ -41,7 +41,7 @@ void set_time(uint32_t epoch_time)
 
 uint32_t uptime_to_epoch(uint32_t uptime_ms)
 {
-	return epoch_time_ref + ((uptime_ms - uptime_ms_ref) / 1000);
+	return epoch_time_ref + ((int32_t)(uptime_ms - uptime_ms_ref) / 1000);
 }
 
 uint32_t time(void)
